Fix Opencast Studio return link for Ilias 5.x
The `getLinkTarget` method in Ilias 5.4 has `xmlStyle` default to `true`
(see [1]). In Ilias 6.0, it defaults to `false` instead (see [2]). We
need this to be `false` as otherwise `&` gets turned into `&amp;` which
Opencast Studio does not decode. To ensure this works with both Ilias
versions, we just specify the parameters explicitly.

[1]: http://ildoc.hrz.uni-giessen.de/ildoc/release_5-4/html/d9/d05/classilCtrl.html#af1e03eb3719c5d2891a8e9381940425d
[2]: http://ildoc.hrz.uni-giessen.de/ildoc/release_6/html/d9/d05/classilCtrl.html#add1a0ceadd022159137cb5ccfb9bd364